put `realpath` into its own module
That way we avoid making `convert` a catch-all and maybe open up the
crate for more refactoring around the existing modules like 'convert'.
